requirements.txt
flask==2.3.3
flask-cors==4.0.0
plaid-python==17.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.1
psycopg2-binary==2.9.9
werkzeug==2.3.7
numpy==1.26.0
openpyxl==3.1.2